This is the penultimate patch for integration of memory and exception manager APIs into pgpool. The commit is very large mainly because of replacing of pool_error(), pool_debug() and pool_log() function calls to ereport(ERROR,..) ereport(DEBUG,..) and ereport(LOG,..) Below is the description of few areas changed by the commit
Some Important Memory Contexts
============================

    ProcessLoopContext:
    This context resets at every main loop iteration of a process

    TopMemoryContext:
    This context is same as PostgreSQL's TopMemoryContext and lives for a life time of process.

    ErrorContext:
    Same as PostgreSQL's ErrorContext and used for error processing.

    QueryContext:
    Used by pgpool child process and it lives for single query cycle.

Lifecycle of Important Objects
==============================

    SessionContext:
    The session context consists of a backend and frontend connection and it lives for a single iteration of main child process loop.
    So the session context object is created in ProcessLoopContext.

    ProcessContext:
    ConnectionPool:
    BackendConnection:
    The above objects are required for the life of process hence are created in TopMemoryContext which lives throughout the process life time.

    Child Frontend connection:
    Frontend connection object is placed in the ProcessLoopContext since new frontend connection is created at every main loop iteration of pgpool child process.

    PCP Frontend connection:
    This object can live till the pcp process so the object for pcp frontend is created in TopMemoryContext.

new configuration parameters
===========================

    The patch also adds three new configuration parameters to control the error reporting style and details.
    Parameters names and their behaviours are borrowed from PostgreSQL.
    int log_error_verbosity;    /* controls how much detail about error should be emitted */
    int client_min_messages;    /*controls which message should be sent to client */
    int log_min_messages;       /*controls which message should be emitted to server log */

Some other notable changes introduced by the patch.
================================================

    The patch changes the error reporting behaviour and now pgpool-II tries to send all the error and log messages to the connected frontend clients (controllable by "client_min_messages" configuration pa
    The patch added an extra call back "on_system_exit" to the elog API, which is not present in the PostgreSQL.

+++ src/watchdog/watchdog.c
@@ -75,16 +75,19 @@ void
 wd_kill_watchdog(int sig)
 {
 	int i;
-
-	kill (lifecheck_pid, sig);
-	kill (child_pid, sig);
+    if(lifecheck_pid > 0)
+        kill (lifecheck_pid, sig);
+    if(child_pid > 0)
+        kill (child_pid, sig);
 
 	if (!strcmp(pool_config->wd_lifecheck_method, MODE_HEARTBEAT))
 	{
 		for (i = 0; i < pool_config->num_hb_if; i++)
 		{
-			kill (hb_receiver_pid[i], sig);
-			kill (hb_sender_pid[i], sig);
+            if(hb_receiver_pid[i] > 0)
+                kill(hb_receiver_pid[i], sig);
+            if(hb_sender_pid[i] > 0)
+                kill (hb_sender_pid[i], sig);
 		}
 	}
 }
@@ -193,6 +196,7 @@ fork_a_lifecheck(int fork_wait_time)
 
 		return pid;
 	}
+    on_exit_reset();
 
 	if (fork_wait_time > 0) {
 		sleep(fork_wait_time);
